add kernel file and a entry-point function to the classifier
